fixes more namespacing, adds files needed for base management back moved to base management folder
