+++ Assets/Scripts/GameData/Equipment/IEquipment.cs
@@ -1,10 +1,7 @@
 ﻿namespace SwordAndBored.GameData.Equipment
 {
-    interface IEquipment : IDescriptable
+    public interface IEquipment : IDescriptable
     {
-        /// <summary>
-        /// The modifiers granted by this equipment
-        /// </summary>
         // TODO Will be used for passive abilities in the future
         //IModifier[] Modifiers { get; }
     }
