+++ Assets/Scripts/GameData/IDescriptable.cs
@@ -7,6 +7,6 @@
     {
         string Name { get; set; }
         string Description { get; set; }
-        string Flavor_Text { get; set; }
+        string FlavorText { get; set; }
     }
 }
