+++ Assets/Scripts/Strategy/BaseManagement/IModelable.cs
@@ -1,9 +1,8 @@
-﻿/*
- * namespace SwordAndBored.GameData
+﻿
+namespace SwordAndBored.StrategyView.BaseManagement
 {
     public interface IModelable
     {
         string ModelName { get; set; }
     }
 }
-*/
