+++ Assets/Scripts/Strategy/BaseManagement/IModelable.cs.meta
