+++ Assets/Scripts/Strategy/BaseManagement/IUpdatable.cs
@@ -1,13 +1,11 @@
-﻿/*
- * using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace SwordAndBored.GameData
+namespace SwordAndBored.StrategyView.BaseManagement
 {
     public interface IUpdatable
     {
         void Update();
     }
 }
-*/
